+++ src/union_find/union_find.h
@@ -6,34 +6,34 @@
 #include <stdint.h>
 
 typedef struct {
-    uint16_t N;     // Number of vertices
-    uint16_t *id;   // id[i] = parent of i
-    uint16_t *size; // size[i] = number of vertices in subtree rooted at i
+    int N;     // Number of vertices
+    int *id;   // id[i] = parent of i
+    int *size; // size[i] = number of vertices in subtree rooted at i
 } UF;
 
 /// @brief Initialize the Union-Find structure
 /// @param N Number of vertices
 /// @return UF* Pointer to the Union-Find structure
-UF* UF_init(uint16_t N);
+UF* UF_init(int N);
 
 /// @brief Find the root of the tree
 /// @param uf Pointer to the Union-Find structure
 /// @param i Vertex
 /// @return uint16_t Root of the tree
-uint16_t UF_find(UF* uf, uint16_t i);
+int UF_find(UF* uf, int i);
 
 /// @brief Union two trees
 /// @param uf Pointer to the Union-Find structure
 /// @param p Vertex
 /// @param q Vertex
-void UF_union(UF* uf, uint16_t p, uint16_t q);
+void UF_union(UF* uf, int p, int q);
 
 /// @brief Check if two vertices are connected
 /// @param uf Pointer to the Union-Find structure
 /// @param p Vertex
 /// @param q Vertex
 /// @return int 1 if connected, 0 otherwise
-int UF_connected(UF* uf, uint16_t p, uint16_t q);
+int UF_connected(UF* uf, int p, int q);
 
 void UF_free(UF* uf);
